Move profiler hook from contrib to core.
PiperOrigin-RevId: 171194291
